refactor : 로또 사는 기능, LottoCashier 분리
